requirements.txt
flask>=3.0
anthropic>=0.40
google-api-python-client>=2.100
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.0
